src/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
whoosh>=2.7.4
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
rich>=13.0.0
scikit-learn>=1.3.0
ollama>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0